Move ES ingest functionality from vectorize to provision_es
